requirements.txt
# Core dependencies
opencv-python~=4.12.0.88
numpy~=2.2.6
Pillow~=11.3.0
flask~=3.1.1
flask-cors

# AI Framework (optional - can be installed separately)
# tensorflow
# torch
# torchvision

# Face Recognition (optional)
# dlib
# face-recognition

# Image Processing
scipy~=1.16.1
scikit-image~=0.25.2

# Utilities
requests~=2.32.5
python-dotenv
torch~=2.8.0
torchvision~=0.23.0